backend/requirements.txt
Django
djangorestframework
psycopg2
django-cors-headers
django-rest-auth
django-allauth
django-registration
dateutil